Remove failed IPCS experiment related to hydrostatic pressure
The momentum equations may not converge due to not having the right
pressure, especially when starting from zero pressure and expecting
hydrostatic pressure after the first time step. The hypotesis was that
rerunning the pressure and momentum inner loop to get a better pressure
might be enough to get convergence. This turned out to not be such a
good idea.

Calculating the hydrostatic pressure at the start of the time step to
avoid missing a large (but known) pressure in the momentum prediction
turned out to work much better