Int. by parts of RHS in IPCS pressure eq improves convergence and divergence
After integration by parts of div(u)*q we can use the same velocity flux in the
pressure Poisson equation as we use in the BDM postprocessing. This improves the
divergence free properties of the solver and also the convergence rate which is
now 3 in L2 on the Kovasznay flow test case for velocities in DG2. Previously
the convergence rate was 2.5 which was of course sub-optimal.