Better timing of post-timestep hooks
Also adds ability to get timings on console in the middle of a running
Ocellaris simulation. This is usefull if you suspect that a hook that
was added is taking a long time and you may want to kill the simulation
to rewrite the hook, but you want to make sure it is actually t he hook
and not an intrinsic part of the solver that is slow.