Implement writing of restart files
Ocellaris can now write both planned restart files (with given interval
from input file) and also when requested on the command line while
running with "r+<ENTER>".

Restarting has not yet been implemented, but the restart files are
probably complete enough to implement this now. Restart files contain
fields such as velocity and pressure along with the contents of the
input dictionary serialised to YAML and the full list of logged messages