Add stop/restart functionality to Ocellaris
The machinery around setting up the simulation and running the solver
has been refactored to simplify the implementation of simulation
restart. The same functionality will hopefully make it easy to use
Ocellaris from a script as well. As an example postprocessing results
from a simulation should now be quite easy by loading a restart file:

    import dolfin
    from ocellaris import Simulation

    sim = Simulation()
    sim.io.load_restart_file_input(h5_file)
    sim.setup()
    sim.io.load_restart_file_results(h5_file)
    
    u = sim.data['u']
    dolfin.plot(u, title='velocity')
    dolfin.interactive()