Small updates + add support for Matrix & Vector file IO
Needed to dump matrices and vectors to file to debug massive slowdown
when updating dolfin. It turns out that PETScLUSolver of type "umfpack"
(the default) on 1 CPU now takes 17-18 seconds vs 3-4 seconds on
previous (2016.2.0.dev0 hashdist install probably from August 2016).
Both of these measured on the same matrix/vector system

Still remains to be seen if it is the Suitesparse version (4.0.2 in
hashdist vs 4.4.3