Support limiting u_conv as well
This is for testing mainly, it should probably never be activated with
the HT limiter since it destroys the solenoidal properties. This option
allows comparison with how solvers normally work where u^n == u_conv^n.
This is no longer the case for the coupled Ocellaris solver (and
probably will change also for the other solvers)