Update to new solenoidal API
Hopefully this will in the end make the code easier to reason about. The
problem with a research code is that since you do not know where you
will end up it is hard to plan a good design. The use of different
limiters for u and u_conv was not planned. It should still be cleanded
up even more if this is the final design, but for now this will
hopefully do the job and untangle some strange misuses of the old API