Timeout context manager.
Not used yet, but planned to be used for aborting stuck simulations.
Some analyses are being stuck at 100% CPU in some kind of infinite loop.
Probably in a compiled extension (PETSc?), but not certain since this is
not deterministic and happens after some hours of running. Not traceback
is given on Ctrl+C, which could indicate a problem that is not in the
Python code.