Failed attempt at speeding up the Krylov solver convergence
In the momentum solver part of IPCS-A, more Krylov iterations are needed
when rho_max = 1000 and rho_min = 1 compared to when rho_min = 10.
This could lead one to believe that by slowly decrasing rho_min towards
the physical value it should be  possible to speed up the convergence.

But, while this iteration procedure works, it does not seem to speed up
the convergence. Where before there was one step
0 1.0 127         (rho_min = 1.0,  number of Krylov iterations=127),
the result after the modification is
0 500 7
1 50 17
2 5 57
3 1.0 125

So, two iterations are saved at the lowest rho_min, and lots are used to
accomplish this. Note the number of iterations for the highest rho_min:
the density ratio = rho_max/rho_min does most definitively affect the
condition number of the matrix