Another attempt at speeding up momentum solve via continuation
Assemble two systems. One "nice" with low condition number needing few
Krylov iterations, and one system which we actually want to solve. The
difference between the systems is just the density in the light (air)
phase. For a number of steps, blend together the systems starting
with the nice system and ending with the original system. Each step runs
a number of Krylov iterations on the blended system.

From perliminary testing this is still slower than the original method
without any such steps :-(